requirements.txt
streamlit: 1.43.0
python-dotenv: 1.0.0
openai: 0.28.0
requests: 2.31.0
python-docx: 1.1.2
pandas: 2.2.0
azure-storage-blob: 12.23.1
xlsxwriter: 3.2.2
statsmodels: 0.14.4
plotly: 6.0.0
matplotlib: 3.7.1
wordcloud: 1.9.4
seaborn: 0.12.2
numpy: 1.24.2
psycopg2: 2.9.10
psycopg2-binary==2.9.9
xlwt==1.3.0
slack-sdk==3.27.1
python-canvasapi==2.3.0
python-docx==0.8.11
python-pptx==0.6.23